refactor: 화면별로 패키지 분리
