+++ app/src/androidTest/java/woowacourse/movie/uiTest/CompletedBookingActivityTest.kt
@@ -8,11 +8,11 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
-import woowacourse.movie.activity.CompleteActivity
+import woowacourse.movie.completedbooking.CompletedBookingActivity
 
-class CompleteActivityTest {
+class CompletedBookingActivityTest {
     @get:Rule
-    val activityRule = ActivityScenarioRule(CompleteActivity::class.java)
+    val activityRule = ActivityScenarioRule(CompletedBookingActivity::class.java)
 
     @Test
     fun haveCancelInfo() {
