+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".activity.ReservationActivity"
+            android:name=".detailbooking.DetailBookingActivity"
             android:exported="false"
-            android:parentActivityName=".activity.MovieListActivity"/>
+            android:parentActivityName=".movielist.MovieListActivity"/>
         <activity
-            android:name=".activity.CompleteActivity"
+            android:name=".completedbooking.CompletedBookingActivity"
             android:exported="false"
-            android:parentActivityName=".activity.MovieListActivity" />
+            android:parentActivityName=".movielist.MovieListActivity" />
         <activity
-            android:name=".activity.MovieListActivity"
+            android:name=".movielist.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
