+++ app/src/main/res/layout/activity_complete.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity.CompleteActivity">
+    tools:context=".completedbooking.CompletedBookingActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
