+++ app/src/main/res/layout/activity_reservation.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ReservationActivity"
+    tools:context=".detailbooking.DetailBookingActivity"
     android:orientation="vertical">
 
     <ImageView
